requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.11
botocore==1.34.11
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.7
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.8
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
django-storages==1.14.2
django-timezone-field==6.1.0
filelock==3.13.1
ghasedakpack==0.1.13
hiredis==2.3.2
identify==2.5.33
idna==3.6
jmespath==1.0.1
kavenegar==1.1.2
kombu==5.3.4
ndg-httpsclient==0.5.1
nodeenv==1.8.0
Pillow==10.1.0
platformdirs==4.1.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pyOpenSSL==23.3.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
s3transfer==0.10.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.0.7
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.12
